Modified og:image tag for product pages to use social media preview
This commit links to the following hack day project
https://hackdays.shopify.io/projects/14631

The theme is able to provide it's own dynamic open graph image by rendering
an alternative view of the product using the view name 'open-graph'.

In the primary view we then point to the screenshots service which takes
an image of that view and returns the resultant image for use as a more
compelling open graph image.

fixes https://github.com/Shopify/hackdays-social-image-previews/issues/6