Adds experimental open-graph product view
This commit relates to the following hack day project
https://hackdays.shopify.io/projects/14631

This view is an exerimental approach that has been validated but
needs finessing. Similar to the default product view it needs a
black based approach to allow configuration.

The tricker aspect is that it must all fit within a fixed width.
Currently set to 1200 x 630 which is a sensible default size for
open graph images.