Upgrade from img_url to image_url
`img_url` has been deprecated. The new filter, `image_url` is more
straightforward to understand and doesn't require the use of the
'master' size specification.

The next version of theme-check will flag their use as warnings, this PR
fixes that before Dawn lights up with 250+ warnings.

And minor theme-check --autocorrect changes that can be fixed with the
JsonSchemaFormat autocorrector.