fix(#51542): wrap dummy form in product form class
The dummy form is generated because we want to use properties from the liquid product from drop. It doesn't serve any actual purpose in the html.

Normally product forms are wrapped in a <product-form> custom element which have some styling applied. We don't want to do that because the custom element also adds a bunch of js that we don't want. Instead we just use a div with the class.