feat(#51543): Correct checkbox bahvior in all cases
When js is disabled the form displays with no checkbox
When js is enabled the form is hidden at page load and the checkbox is unhidden and enabled

This works well but is susceptible to content flashes as the form is first shown then quickly hidden. Probably unacceptable.

An obvious solution would be to check the checkbox by default which would mean there is no need to hide the form on load.

A more extreme solution is to use a noscript tag to completely sever the link between the two sets of markup. I don't like this solution as it bloats the source but it does neatly take care of the content flash