feat(#51542): Error formatting for no js recipient form
When we read errors out of the flash state we want them displayed nicely inline on the recipient form.

We can do this by ripping off the pattern frolm customer registration. There are a few styles which have been implemented as
specific .customer styles which should really be a generic class but not a huge issue to just use the customer class in our form as well