This attempts to standardise error handling between liquid and javascript
Liquid errors already render nicely next to the fields they belong to. We'd eventually like the same behaviour in JS. That will require tweaking the backend error payload a bit but is otherwise achievable.

We first have to disable the existing product form error handling as this appears next to the buy buttons

Instead we publish the error response on the pub/sub bus and display the error at the top of the recipient form.

We have liquid render the error container so it looks roughly the same regardless of whether the error is rendered by liquid or JS.

Next step will be to break the error message down into fields on the backend and unpack the messages on the FE.

Co-authored-by: Frederic Ma <frederic.ma@shopify.com>