Add aria-labelledby attributes to form
This is still not perfect, voiceover on mac will read out:

{aria-labelledby} {placeholder}, edit text {label}

e.g

Recipient's Email Email, edit text Email

I'm not sure it should use the placeholder like that. Ultimately the issue is that we use two labels for the field, perhaps we should scrap that approach and just use one label?