Add aria-expanded and aria-control attribbutes to gift card checkbox
The intent here is to give context to screen readers that the gift card recipient form is being hidden and shown. I'm using the example here https://www.w3.org/WAI/ARIA/apg/patterns/combobox/examples/combobox-autocomplete-list/ as a reference but I'm not getting great results from mac voiceover. Android talkback is better, when I highlight the checkbox it correctly announces the collapsed/expanded state.

I should note that it is only the checkbox that gets the state not the label which would be more commonly selected as it's bigger. I think this is a good first step to solving this issue, though we may want to do more.