Only render media gallery if there is media
I have a downloadable product in my store that doesn't have any media attached to it. The Featured Product section renders the `product-media-gallery` with the product's media.

When a product has no media, it renders the layout just fine—but then has a `1/0` badge at the top to show that it's rendering the first out of no images 🤦 

This little patch makes it so that the product-media-gallery respects products without any media.